Fixup e2e tests.
The shiro auth scheme crucially relies on the pojo intercepted
layer being able to throw an UnauthenticatedException (RTE) up
through the thrift layer to the HTTP servlet filter layer.  The Swift
layer is more robust to catching exceptions and translating them
to thrift protocol-layer TExceptions which thwarted this.  This fix
adds workarounds that leverage Error as the transportation layer
as opposed to RuntimeException.