1st cut at testing codegen.
This uses the compiler to compiler the codegen as a sanity
check and then loads the compiled code to further check properties
of the generated code.  This is a bit awkward given that the loaded
Class is opaque.  A follow up may switch to using compiler-testing
for a slightly less awkward though more tightly coupled whitebox test.